Added player, stength, display classe
Change of plan for the main game. decided to make a game where player
gets to move on x by y map with enemies on it and player get to hit them
with a strength bar by hitting enter on the perfect spot. Player has to
move on to the same tile as the enemy in order to hit them.